+++ move.c
@@ -15,16 +15,16 @@
 static int Move(char *to_path, char *from_path);
 
 
-int MoveFile(FileEntry *fe_ptr, 
-	     unsigned char confirm, 
-	     char *to_file, 
-	     DirEntry *dest_dir_entry, 
+int MoveFile(FileEntry *fe_ptr,
+	     unsigned char confirm,
+	     char *to_file,
+	     DirEntry *dest_dir_entry,
 	     char *to_dir_path,
 	     FileEntry **new_fe_ptr
 	    )
 {
   DirEntry    *de_ptr;
-  LONGLONG    file_size;
+  long long file_size;
   char        from_path[PATH_LENGTH+1];
   char        to_path[PATH_LENGTH+1];
   FileEntry   *dest_file_entry;
@@ -45,7 +45,7 @@ int MoveFile(FileEntry *fe_ptr,
   (void) strcat( to_path, FILE_SEPARATOR_STRING );
   (void) strcat( to_path, to_file );
 
-  
+
   if( !strcmp( to_path, from_path ) )
   {
     MESSAGE( "Can't move file into itself" );
@@ -55,22 +55,22 @@ int MoveFile(FileEntry *fe_ptr,
 
   if( access( from_path, W_OK ) )
   {
-    (void) sprintf( message, 
-		    "Unmoveable file*\"%s\"*%s", 
-		    from_path, 
-		    strerror(errno) 
+    (void) sprintf( message,
+		    "Unmoveable file*\"%s\"*%s",
+		    from_path,
+		    strerror(errno)
 		  );
     MESSAGE( message );
     ESCAPE;
   }
-  
+
 
   if( dest_dir_entry )
   {
     /* Ziel befindet sich im Sub-Tree */
     /*--------------------------------*/
-  
-    (void) GetFileEntry( dest_dir_entry, to_file, &dest_file_entry ); 
+
+    (void) GetFileEntry( dest_dir_entry, to_file, &dest_file_entry );
 
     if( dest_file_entry )
     {
@@ -80,7 +80,7 @@ int MoveFile(FileEntry *fe_ptr,
       if( confirm )
       {
 	term = InputChoise( "file exist; overwrite (Y/N) ? ", "YN\033" );
-    
+
         if( term != 'Y' ) {
 	  result = (term == 'N' ) ? 0 : -1;  /* Abort on escape */
 	  ESCAPE;
@@ -103,26 +103,26 @@ int MoveFile(FileEntry *fe_ptr,
       if( confirm )
       {
 	term = InputChoise( "file exist; overwrite (Y/N) ? ", "YN\033" );
-      
+
         if( term != 'Y' ) {
 	  result = (term == 'N' ) ? 0 : -1;  /* Abort on escape */
 	  ESCAPE;
 	}
       }
-        
+
       if( unlink( to_path ) )
       {
-        (void) sprintf( message, 
-		        "Can't unlink*\"%s\"*%s", 
-		        to_path, 
-		        strerror(errno) 
+        (void) sprintf( message,
+		        "Can't unlink*\"%s\"*%s",
+		        to_path,
+		        strerror(errno)
 		      );
         MESSAGE( message );
         ESCAPE;
       }
     }
   }
-    
+
 
   if( !Move( to_path, from_path ) )
   {
@@ -134,7 +134,7 @@ int MoveFile(FileEntry *fe_ptr,
 
     (void) RemoveFile( fe_ptr );
 
-    
+
     if( dest_dir_entry )
     {
       if( STAT_( to_path, &stat_struct ) )
@@ -142,9 +142,9 @@ int MoveFile(FileEntry *fe_ptr,
         ERROR_MSG( "Stat Failed*ABORT" );
         exit( 1 );
       }
-  
+
       file_size = stat_struct.st_size;
-    
+
       dest_dir_entry->total_bytes += file_size;
       dest_dir_entry->total_files++;
       statistic.disk_total_bytes += file_size;
@@ -157,21 +157,21 @@ int MoveFile(FileEntry *fe_ptr,
       /* File eintragen */
       /*----------------*/
 
-      if( ( fen_ptr = (FileEntry *) malloc( sizeof( FileEntry ) + 
-					    strlen( to_file ) 
+      if( ( fen_ptr = (FileEntry *) malloc( sizeof( FileEntry ) +
+					    strlen( to_file )
 					  ) ) == NULL )
       {
         ERROR_MSG( "Malloc Failed*ABORT" );
         exit( 1 );
       }
-  	
+
       (void) strcpy( fen_ptr->name, to_file );
-        
-      (void) memcpy( &fen_ptr->stat_struct, 
+
+      (void) memcpy( &fen_ptr->stat_struct,
 		     &stat_struct,
 		     sizeof( stat_struct )
 		   );
-  
+
       fen_ptr->dir_entry   = dest_dir_entry;
       fen_ptr->tagged      = FALSE;
       fen_ptr->matching    = Match( fen_ptr->name );
@@ -188,7 +188,7 @@ int MoveFile(FileEntry *fe_ptr,
   }
 
 FNC_XIT:
- 
+
   move( LINES - 3, 1 ); clrtoeol();
   move( LINES - 2, 1 ); clrtoeol();
   move( LINES - 1, 1 ); clrtoeol();
@@ -243,11 +243,11 @@ static int Move(char *to_path, char *from_path)
 
   if( link( from_path, to_path ) )
   {
-    (void) sprintf( message, 
-		    "Can't link \"%s\"*to \"%s\"*%s", 
-		    from_path, 
-		    to_path, 
-		    strerror(errno) 
+    (void) sprintf( message,
+		    "Can't link \"%s\"*to \"%s\"*%s",
+		    from_path,
+		    to_path,
+		    strerror(errno)
 		  );
     MESSAGE( message );
     return( -1 );
@@ -255,10 +255,10 @@ static int Move(char *to_path, char *from_path)
 
   if( unlink( from_path ) )
   {
-    (void) sprintf( message, 
-		    "Can't unlink*\"%s\"*%s", 
-		    from_path, 
-		    strerror(errno) 
+    (void) sprintf( message,
+		    "Can't unlink*\"%s\"*%s",
+		    from_path,
+		    strerror(errno)
 		  );
     MESSAGE( message );
     return( -1 );
@@ -271,18 +271,18 @@ static int Move(char *to_path, char *from_path)
 
 
 
-  
+
 int MoveTaggedFiles(FileEntry *fe_ptr, WalkingPackage *walking_package)
 {
   int  result = -1;
   char new_name[PATH_LENGTH+1];
 
 
-  if( BuildFilename( fe_ptr->name, 
+  if( BuildFilename( fe_ptr->name,
                      walking_package->function_data.mv.to_file,
 		     new_name
 		   ) == 0 )
-  
+
   {
     if( *new_name == '\0' )
     {
@@ -290,7 +290,7 @@ int MoveTaggedFiles(FileEntry *fe_ptr, WalkingPackage *walking_package)
     }
     else
     {
-      result = MoveFile( fe_ptr, 
+      result = MoveFile( fe_ptr,
 		         walking_package->function_data.mv.confirm,
 		         new_name,
 		         walking_package->function_data.mv.dest_dir_entry,
